style: 질문지 카드와 스터디 채널 카드 디자인 통일
