+++ frontend/src/pages/channels/ChannelCard.tsx
@@ -26,22 +26,18 @@ const ChannelCard = ({
         ${inProgress === false ? "hover:shadow-16 hover:ring-1 hover:ring-green-200" : ""}`}
     >
       <div className="flex flex-col justify-between">
-        <div className={"flex-grow flex flex-col items-start"}>
-          <span
-            className={
-              "text-semibold-s text-green-600 bg-green-50 border-custom-s border-gray-300 rounded-2xl py-px px-2"
-            }
-          >
+        <div className="flex-grow flex flex-col items-start">
+          <span className="text-semibold-s text-green-600 bg-green-50 border-custom-s border-gray-300 rounded-2xl py-px px-2">
             {category[0]}
           </span>
           <div className="px-0.5">
-            <h3 className={"text-semibold-m mt-2 mb-0.5"}>{title}</h3>
-            <p className={"text-medium-m text-gray-400"}>
+            <h3 className="text-semibold-m mt-2 mb-0.5">{title}</h3>
+            <p className="text-medium-m text-gray-400">
               {questionListTitle ?? "함께 면접 스터디에 참여해보세요!"}
             </p>
-            <div className={"absolute bottom-5 left-6 text-medium-r flex flex-col"}>
+            <div className="absolute bottom-5 left-6 text-medium-r flex flex-col">
               <span className="text-gray-600">{host}</span>
-              <span className={"text-gray-black flex gap-2 items-center"}>
+              <span className="text-gray-black flex gap-2 items-center">
                 <FaUserGroup className="text-green-400" />
                 참여자 {participants}/{maxParticipants}명
               </span>
