+++ frontend/src/pages/questions/QuestionsList.tsx
@@ -1,4 +1,4 @@
-import QuestionsPreviewCard from "@components/questions/QuestionsPreviewCard.tsx";
+import QuestionsCard from "@/pages/questions/QuestionCard";
 import type { QuestionList } from "./types/QuestionList";
 
 interface QuestionListProps {
@@ -7,23 +7,21 @@ interface QuestionListProps {
 }
 
 const QuestionsList = ({
-  questionList,
-  tab,
+  questionList
 }: QuestionListProps) => {
   if (questionList.length === 0) {
     return <div className="text-center mt-12 text-medium-l text-gray-500">스크랩한 질문지가 없습니다.</div>;
   }
 
   return (
-    <div className="grid grid-cols-1 lg:grid-cols-2 xl:grid-cols-3 gap-4">
+    <div className="w-full grid grid-cols-2 lg:grid-cols-3 gap-4">
       {questionList.map((list) => (
-        <QuestionsPreviewCard
+        <QuestionsCard
           key={list.id}
           id={list.id}
           questionCount={list.questionCount ?? 0}
           category={list.categoryNames ? list.categoryNames[0] : "미분류"}
           title={list.title}
-          isStarred={tab === "SCRAP" ? true : list.isStarred}
           usage={list.usage}
         />
       ))}
